requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-sslcommerz
djangorestframework
djangorestframework-simplejwt
EasyProcess
entrypoint2
filelock
idna
Markdown
MouseInfo
mss
numpy
opencv-contrib-python
pillow
platformdirs
psycopg2-binary
PyAutoGUI
PyGetWindow
PyJWT
PyMsgBox
pyobjc
pyobjc-core
pyobjc-framework-Accessibility
pyobjc-framework-Accounts
pyobjc-framework-AddressBook
pyobjc-framework-AdServices
pyobjc-framework-AdSupport
pyobjc-framework-AppleScriptKit
pyobjc-framework-AppleScriptObjC
pyobjc-framework-ApplicationServices
pyobjc-framework-AppTrackingTransparency
pyobjc-framework-AudioVideoBridging
pyobjc-framework-AuthenticationServices
pyobjc-framework-AutomaticAssessmentConfiguration
pyobjc-framework-Automator
pyobjc-framework-AVFoundation
pyobjc-framework-AVKit
pyobjc-framework-AVRouting
pyobjc-framework-BackgroundAssets
pyobjc-framework-BrowserEngineKit
pyobjc-framework-BusinessChat
pyobjc-framework-CalendarStore
pyobjc-framework-CallKit
pyobjc-framework-CFNetwork
pyobjc-framework-Cinematic
pyobjc-framework-ClassKit
pyobjc-framework-CloudKit
pyobjc-framework-Cocoa
pyobjc-framework-Collaboration
pyobjc-framework-ColorSync
pyobjc-framework-Contacts
pyobjc-framework-ContactsUI
pyobjc-framework-CoreAudio
pyobjc-framework-CoreAudioKit
pyobjc-framework-CoreBluetooth
pyobjc-framework-CoreData
pyobjc-framework-CoreHaptics
pyobjc-framework-CoreLocation
pyobjc-framework-CoreMedia
pyobjc-framework-CoreMediaIO
pyobjc-framework-CoreMIDI
pyobjc-framework-CoreML
pyobjc-framework-CoreMotion
pyobjc-framework-CoreServices
pyobjc-framework-CoreSpotlight
pyobjc-framework-CoreText
pyobjc-framework-CoreWLAN
pyobjc-framework-CryptoTokenKit
pyobjc-framework-DataDetection
pyobjc-framework-DeviceCheck
pyobjc-framework-DictionaryServices
pyobjc-framework-DiscRecording
pyobjc-framework-DiscRecordingUI
pyobjc-framework-DiskArbitration
pyobjc-framework-DVDPlayback
pyobjc-framework-EventKit
pyobjc-framework-ExceptionHandling
pyobjc-framework-ExecutionPolicy
pyobjc-framework-ExtensionKit
pyobjc-framework-ExternalAccessory
pyobjc-framework-FileProvider
pyobjc-framework-FileProviderUI
pyobjc-framework-FinderSync
pyobjc-framework-FSEvents
pyobjc-framework-GameCenter
pyobjc-framework-GameController
pyobjc-framework-GameKit
pyobjc-framework-GameplayKit
pyobjc-framework-HealthKit
pyobjc-framework-ImageCaptureCore
pyobjc-framework-InputMethodKit
pyobjc-framework-InstallerPlugins
pyobjc-framework-InstantMessage
pyobjc-framework-Intents
pyobjc-framework-IntentsUI
pyobjc-framework-IOBluetooth
pyobjc-framework-IOBluetoothUI
pyobjc-framework-IOSurface
pyobjc-framework-iTunesLibrary
pyobjc-framework-KernelManagement
pyobjc-framework-LatentSemanticMapping
pyobjc-framework-LaunchServices
pyobjc-framework-libdispatch
pyobjc-framework-libxpc
pyobjc-framework-LinkPresentation
pyobjc-framework-LocalAuthentication
pyobjc-framework-LocalAuthenticationEmbeddedUI
pyobjc-framework-MailKit
pyobjc-framework-MapKit
pyobjc-framework-MediaAccessibility
pyobjc-framework-MediaLibrary
pyobjc-framework-MediaPlayer
pyobjc-framework-MediaToolbox
pyobjc-framework-Metal
pyobjc-framework-MetalFX
pyobjc-framework-MetalKit
pyobjc-framework-MetalPerformanceShaders
pyobjc-framework-MetalPerformanceShadersGraph
pyobjc-framework-MetricKit
pyobjc-framework-MLCompute
pyobjc-framework-ModelIO
pyobjc-framework-MultipeerConnectivity
pyobjc-framework-NaturalLanguage
pyobjc-framework-NetFS
pyobjc-framework-Network
pyobjc-framework-NetworkExtension
pyobjc-framework-NotificationCenter
pyobjc-framework-OpenDirectory
pyobjc-framework-OSAKit
pyobjc-framework-OSLog
pyobjc-framework-PassKit
pyobjc-framework-PencilKit
pyobjc-framework-PHASE
pyobjc-framework-Photos
pyobjc-framework-PhotosUI
pyobjc-framework-PreferencePanes
pyobjc-framework-PushKit
pyobjc-framework-Quartz
pyobjc-framework-QuickLookThumbnailing
pyobjc-framework-ReplayKit
pyobjc-framework-SafariServices
pyobjc-framework-SafetyKit
pyobjc-framework-SceneKit
pyobjc-framework-ScreenCaptureKit
pyobjc-framework-ScreenSaver
pyobjc-framework-ScreenTime
pyobjc-framework-ScriptingBridge
pyobjc-framework-SearchKit
pyobjc-framework-Security
pyobjc-framework-SecurityFoundation
pyobjc-framework-SecurityInterface
pyobjc-framework-SensitiveContentAnalysis
pyobjc-framework-ServiceManagement
pyobjc-framework-SharedWithYou
pyobjc-framework-SharedWithYouCore
pyobjc-framework-ShazamKit
pyobjc-framework-Social
pyobjc-framework-SoundAnalysis
pyobjc-framework-Speech
pyobjc-framework-SpriteKit
pyobjc-framework-StoreKit
pyobjc-framework-Symbols
pyobjc-framework-SyncServices
pyobjc-framework-SystemConfiguration
pyobjc-framework-SystemExtensions
pyobjc-framework-ThreadNetwork
pyobjc-framework-UniformTypeIdentifiers
pyobjc-framework-UserNotifications
pyobjc-framework-UserNotificationsUI
pyobjc-framework-VideoSubscriberAccount
pyobjc-framework-VideoToolbox
pyobjc-framework-Virtualization
pyobjc-framework-Vision
pyobjc-framework-WebKit
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
requests
rubicon-objc
setuptools
sqlparse
sslcommerz-lib
typing_extensions
urllib3
virtualenv
wheel
